Monster filtering tech pass
To get this to work I set it up so that each level "unlocks" classes/quirks (if you don't want a quirk gained in a level just leave it null.   Haven't gotten the UI direction just yet so it's currently pretty ugly, definitely not what it'll look like in the end.